requirements.txt
Flask
google-generativeai
Pillow
flask-cors
python-dotenv
gunicorn